requirements.txt
boto3==1.9.139
botocore==1.12.139
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.3
django-ckeditor==5.6.1
django-cors-headers==3.0.2
django-embed-video==1.2.0
django-js-asset==1.2.2
django-storages==1.7.1
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.22.0
rest-framework-generic-relations==1.2.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.24.3
